requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
psycopg2-binary==2.9.4
pytz==2022.5
sqlparse==0.4.2
gunicorn==20.1.0
django-rest-knox==4.2.0
django-cors-headers==3.13.0
Pillow==9.3.0
pandas==1.5.2
openpyxl==3.0.10